Created tests folder and moved tests there. Moved enzyme config to src
