+++ src/__tests__/Nav.test.js
@@ -4,7 +4,7 @@ import "@testing-library/jest-dom/extend-expect"
 import { render } from "@testing-library/react"
 import { MemoryRouter } from "react-router-dom"
 
-import Nav from "./Nav"
+import Nav from "../components/Nav"
 
 describe("NavBar", () => {
   let component
