+++ src/__tests__/WizardStatusMessageHandler.test.js
@@ -3,9 +3,8 @@ import React from "react"
 import { Provider } from "react-redux"
 import configureStore from "redux-mock-store"
 
-import { shallow, mount } from "../../../../enzyme"
-
-import WizardStatusMessageHandler from "./WizardStatusMessageHandler"
+import WizardStatusMessageHandler from "../components/NewDraftWizard/WizardForms/WizardStatusMessageHandler"
+import { shallow, mount } from "../enzyme"
 
 const mockStore = configureStore([])
